+++ docs/source/cai.api.rst
@@ -1,0 +1,50 @@
+cai.api package
+===============
+
+.. automodule:: cai.api
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+cai.api.client module
+---------------------
+
+.. automodule:: cai.api.client
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cai.api.flow module
+-------------------
+
+.. automodule:: cai.api.flow
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cai.api.friend module
+---------------------
+
+.. automodule:: cai.api.friend
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cai.api.group module
+--------------------
+
+.. automodule:: cai.api.group
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+cai.api.login module
+--------------------
+
+.. automodule:: cai.api.login
+   :members:
+   :undoc-members:
+   :show-inheritance:
